Improvements and refactory
