+++ commands/customcommands/listCustomCommands.js
@@ -1,8 +1,9 @@
 const { Command } = require('discord.js-commando');
 const logger = require('heroku-logger');
 const { RichEmbed } = require('discord.js');
-const mongoConnection = require('../../modules/mongoConnection.js');
-const errorMessage = require('../../modules/errorMessage.js');
+
+const mongoConnection = require('../../modules/connection/mongoConnection.js');
+const errorMessage = require('../../modules/message/errorMessage.js');
 
 const logName = '[GetCustomCommand]';
 const wrapLine = '\n';
