+++ modules/gateway/searchSystemFactionFromEdsm.js
