+++ modules/gateway/searchWingInfosFromEddb.js
