+++ modules/service/normalizeWingInfoFromEddb.js
